Update www/py-gunicorn to 19.1.0
pkgsrc changes:
---------------
- Cleanups

Upstream changes:
-----------------

Complete changelog in share/doc/py-gunicorn/2014-news.rst.

19.1
====

Bugfix release.

19.0
====
Gunicorn 19.0 is a major release with new features and fixes. This
version improve a lot the usage of Gunicorn with python 3 by adding two
new workers to it: `gthread` a fully threaded async worker using futures
and `gaiohttp` a worker using asyncio.

Breaking Changes
~~~~~~~~~~~~~~~~
Switch QUIT and TERM signals
++++++++++++++++++++++++++++
With this change, when gunicorn receives a QUIT all the workers are
killed immediately and exit and TERM is used for the graceful shutdown.
Note: the old behaviour was based on the NGINX but the new one is more
correct according the following doc:
https://www.gnu.org/software/libc/manual/html_node/Termination-Signals.html
also it is complying with the way the signals are sent by heroku:
https://devcenter.heroku.com/articles/python-faq#what-constraints-exist-when-developing-applications-on-heroku

Deprecations
+++++++++++++
`run_gunicorn`, `gunicorn_django` and `gunicorn_paster` are now
completely deprecated and will be removed in the next release. Use the
`gunicorn` command instead.